requirements.txt
flask==2.1.0
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
werkzeug==2.1.1
gunicorn==20.1.0
email-validator==1.1.3
pytest==7.0.1
pytest-flask==1.2.0 